feat: move the new logic to the webdriver-image-comparison module
