+++ packages/webdriver-image-comparison/src/helpers/constants.ts
@@ -1,6 +1,7 @@
 import type { IosOffsets } from './constants.interfaces.js'
 import type { ResizeDimensions } from '../methods/images.interfaces.js'
-import type { TestContext } from 'src/commands/check.interfaces.js'
+import type { TestContext } from '../commands/check.interfaces.js'
+import type { DeviceRectangles } from '../methods/instanceData.interfaces.js'
 
 export const DEFAULT_FORMAT_STRING = '{tag}-{browserName}-{width}x{height}-dpr-{dpr}'
 export const STORYBOOK_FORMAT_STRING = '{tag}-{logName}-{width}x{height}-dpr-{dpr}'
@@ -46,6 +47,13 @@ export const DEFAULT_TABBABLE_OPTIONS = {
         color: '#000',
         width: 1,
     },
+}
+export const DEVICE_RECTANGLES: DeviceRectangles = {
+    statusBarAndAddressBar: { top: 0, left: 0, width: 0, height: 0 },
+    viewport: { top: 0, left: 0, width: 0, height: 0 },
+    bottomBar: { top: 0, left: 0, width: 0, height: 0 },
+    leftSidePadding: { top: 0, left: 0, width: 0, height: 0 },
+    rightSidePadding: { top: 0, left: 0, width: 0, height: 0 },
 }
 
 /**
