+++ packages/webdriver-image-comparison/src/helpers/utils.ts
@@ -9,6 +9,10 @@ import type {
     GetToolBarShadowPaddingOptions,
     ScreenshotSize,
 } from './utils.interfaces.js'
+import type { DeviceRectangleBound, DeviceRectangles } from '../methods/instanceData.interfaces.js'
+import type { Executor } from '../methods/methods.interfaces.js'
+import { checkMetaTag } from '../clientSideScripts/checkMetaTag.js'
+import { injectWebviewOverlay } from '../clientSideScripts/injectWebviewOverlay.js'
 
 /**
  * Get and create a folder
@@ -324,4 +328,195 @@ export function isStorybook(){
  */
 export function updateVisualBaseline(): boolean {
     return process.argv.includes('--update-visual-baseline')
+}
+
+/**
+ * Get the mobile screen size, this is different for native and webview
+ */
+export async function getMobileScreenSize({
+    executor,
+    isIOS
+}: {
+    executor: Executor,
+    isIOS: boolean
+}): Promise<{ height: number; width: number }> {
+    let height = 0, width = 0
+
+    if (isIOS) {
+        ({ screenSize: { height, width } } = (await executor('mobile: deviceScreenInfo')) as {
+            statusBarSize: { width: number, height: number },
+            scale: number,
+            screenSize: { width: number, height: number },
+        })
+    // It's Android
+    } else {
+        const { realDisplaySize } = (await executor('mobile: deviceInfo')) as { realDisplaySize: string }
+
+        if (!realDisplaySize || !/^\d+x\d+$/.test(realDisplaySize)) {
+            throw new Error(`Invalid realDisplaySize format. Expected 'widthxheight', got "${realDisplaySize}"`)
+        }
+        [width, height] = realDisplaySize.split('x').map(Number)
+    }
+
+    return { height, width }
+}
+
+/**
+ * Load a base64 HTML page in the browser
+ */
+export async function loadBase64Html({ executor, isIOS, url }: {executor:Executor, isIOS:boolean, url:any}): Promise<void> {
+    const htmlContent = `
+        <html>
+        <head>
+            <title>Base64 Page</title>
+            <meta name="viewport" content="width=device-width, initial-scale=1">
+            <script>
+                document.addEventListener("DOMContentLoaded", function() {
+                    // Force correct viewport settings
+                    const meta = document.querySelector("meta[name='viewport']");
+                    if (!meta) {
+                        const newMeta = document.createElement("meta");
+                        newMeta.name = "viewport";
+                        newMeta.content = "width=device-width, initial-scale=1";
+                        document.head.appendChild(newMeta);
+                    }
+                });
+            </script>
+        </head>
+        <body>
+            <h1>Hello from Base64!</h1>
+            <p>This page was loaded without visiting a URL.</p>
+        </body>
+        </html>`
+
+    const base64Html = Buffer.from(htmlContent).toString('base64')
+
+    await url(`data:text/html;base64,${base64Html}`)
+
+    if (isIOS) {
+        await executor(checkMetaTag)
+    }
+}
+
+/**
+ * Execute a native click
+ */
+export async function executeNativeClick({ executor, isIOS, x, y }:{executor: Executor, isIOS:boolean, x: number, y: number}): Promise<void> {
+    if (isIOS) {
+        return executor('mobile: tap', { x, y })
+    }
+
+    try {
+        // The `clickGesture` is not working on Appium 1, only on Appium 2
+        await executor('mobile: clickGesture', { x, y })
+    } catch (error: unknown) {
+        if (
+            error instanceof Error &&
+          /WebDriverError: Unknown mobile command.*?(clickGesture|tap)/i.test(error.message)
+        ) {
+            console.log(
+                'Error executing `clickGesture`, falling back to `doubleClickGesture`. This likely means you are using Appium 1. Is this intentional?'
+            )
+            await executor('mobile: doubleClickGesture', { x, y })
+        } else {
+            throw error
+        }
+    }
+}
+
+/**
+ * Get the webview click and viewport dimensions
+ */
+async function getMobileWebviewClickAndDimensions(executor:Executor): Promise<DeviceRectangleBound> {
+    return executor(() => {
+        const overlay = document.querySelector('[data-test="ics-overlay"]') as HTMLElement | null
+        const defaultValue = { top: 0, left: 0, width: 0, height: 0 }
+
+        if (!overlay || !overlay.dataset.icsWebviewData) {
+            return defaultValue
+        }
+
+        overlay.remove()
+
+        try {
+            return JSON.parse(overlay.dataset.icsWebviewData)
+        } catch {
+            return defaultValue
+        }
+    })
+}
+
+/**
+ * Get the mobile viewport position, we determine this by:
+ * 1. Loading a base64 HTML page
+ * 2. Injecting an overlay on top of the webview with an event listener that stores the click position in the webview
+ * 3. Clicking on the overlay in the center of the screen with a native click
+ * 4. Getting the data from the overlay and removing it
+ * 5. Calculating the position of the viewport based on the click position of the native click vs the overlay
+ * 6. Returning the calculated values
+ */
+export async function getMobileViewPortPosition({
+    initialDeviceRectangles,
+    isAndroid,
+    isIOS,
+    isNativeContext,
+    methods: {
+        executor,
+        getUrl,
+        url,
+    },
+    nativeWebScreenshot,
+    screenHeight,
+    screenWidth,
+}: {
+    initialDeviceRectangles: DeviceRectangles,
+    isNativeContext: boolean,
+    isAndroid: boolean,
+        isIOS: boolean,
+        methods: {
+            executor: Executor,
+            getUrl: () => Promise<string>,
+            url: (arg:string) => Promise<WebdriverIO.Request|void>,
+    }
+    nativeWebScreenshot: boolean,
+    screenHeight: number,
+    screenWidth: number,
+}): Promise<DeviceRectangles> {
+
+    if (!isNativeContext && (isIOS || (isAndroid && nativeWebScreenshot))) {
+        const currentUrl = await getUrl()
+        // 1. Load a base64 HTML page
+        await loadBase64Html({ executor, isIOS, url })
+        // 2. Inject an overlay on top of the webview with an event listener that stores the click position in the webview
+        await executor(injectWebviewOverlay, isAndroid)
+        // 3. Click on the overlay in the center of the screen with a native click
+        const nativeClickX = screenWidth / 2
+        const nativeClickY = screenHeight / 2
+        await executeNativeClick({ executor, isIOS, x: nativeClickX, y: nativeClickY })
+        // We need to wait a bit here, otherwise the click is not registered
+        await waitFor(100)
+        // 4a. Get the data from the overlay and remove it
+        const { top, left, width, height } = await getMobileWebviewClickAndDimensions(executor)
+        // 4.b reset the url
+        await url(currentUrl)
+        // 5. Calculate the position of the viewport based on the click position of the native click vs the overlay
+        const viewportTop = Math.max(0, Math.round(nativeClickY - top))
+        const viewportLeft = Math.max(0, Math.round(nativeClickX - left))
+        const statusBarAndAddressBarHeight = Math.max(0, Math.round(viewportTop))
+        const bottomBarHeight = Math.max(0, Math.round(screenHeight - (viewportTop + height)))
+        const leftSidePaddingWidth = Math.max(0, Math.round(viewportLeft))
+        const rightSidePaddingWidth = Math.max(0, Math.round(screenWidth - (viewportLeft + width)))
+        const deviceRectangles = {
+            statusBarAndAddressBar: { top: 0, left: 0, width: screenWidth, height: statusBarAndAddressBarHeight },
+            viewport: { top: viewportTop, left: viewportLeft, width: width, height: height },
+            bottomBar: { top: viewportTop + height, left: 0, width: screenWidth, height: bottomBarHeight },
+            leftSidePadding: { top: viewportTop, left: 0, width: leftSidePaddingWidth, height: height },
+            rightSidePadding: { top: viewportTop, left: viewportLeft + width, width: rightSidePaddingWidth, height: height },
+        }
+
+        return deviceRectangles
+    }
+
+    // No WebView detected, return empty values
+    return initialDeviceRectangles
 }
