+++ packages/webdriver-image-comparison/src/index.ts
@@ -9,7 +9,8 @@ import checkFullPageScreen from './commands/checkFullPageScreen.js'
 import checkTabbablePage from './commands/checkTabbablePage.js'
 import { ClassOptions } from './helpers/options.interfaces.js'
 import { ImageCompareResult } from './methods/images.interfaces.js'
-import { IOS_OFFSETS, FOLDERS, DEFAULT_TEST_CONTEXT } from './helpers/constants.js'
+import { DEFAULT_TEST_CONTEXT, IOS_OFFSETS, FOLDERS, NOT_KNOWN } from './helpers/constants.js'
+import { getMobileScreenSize, getMobileViewPortPosition } from './helpers/utils.js'
 
 export type { ScreenshotOutput } from './helpers/afterScreenshot.interfaces.js'
 export type {
@@ -37,6 +38,7 @@ export {
     DEFAULT_TEST_CONTEXT,
     IOS_OFFSETS,
     FOLDERS,
+    NOT_KNOWN,
     saveScreen,
     saveElement,
     saveFullPageScreen,
@@ -45,4 +47,6 @@ export {
     checkElement,
     checkFullPageScreen,
     checkTabbablePage,
+    getMobileScreenSize,
+    getMobileViewPortPosition,
 }
